chore!: bump primer backend, adapt to new selection info api
When selecting a non-empty hole, the API now emits information about
both the expected and actual types. We adapt to this change by ignoring
the actual type and just displaying the expected one (this matches the
behaviour with the old backend version). In a subsequent commit we will
show both.

BREAKING CHANGE: the API is different from the previous backend pin. Thus
this version of the frontend cannot talk to the previous version of the
backend and vice versa.

Signed-off-by: Ben Price <ben@hackworthltd.com>