chore: Bump Primer pin
This backend version outputs flavors for edges. Here we make the minimum changes in order to get this codebase to compile. Making use of the flavors in order to provide visual labels on edges is left for future work.

Note that the examples in `trees.ts` have been updated so that they typecheck, but rather than laboriously choosing the correct edge flavors we just use `"Hole"` everywhere.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>