chore(nodejs): upgrade to React Query v5
This version has some fairly major changes, but most of it is hidden
from us thanks to Orval. We have made the following notable changes,
however:

* `cacheTime` has been renamed to `gcTime`. As far as I can tell,
there are no behavioral changes and this is simply a rename to reduce
confusion. Therefore, this change should be purely cosmetic. (See
https://github.com/TanStack/query/issues/4678,
https://github.com/TanStack/query/discussions/1217, and
https://github.com/TanStack/query/pull/4829 for reference.)

* Use object-style parameters to query methods where necessary. These
changes should be purely cosmetic.

* React Query v5 is written in TypeScript, so we can take advantage of
better types. The one place we currently do that is in the `Edit`
component's `useGetProgram` return types. Note that we also use
`isPending` rather than `isLoading`. The docs are a bit confusing on
this point as to whether they're equivalent; see
https://tanstack.com/query/latest/docs/framework/react/guides/migrating-to-v5#status-loading-has-been-changed-to-status-pending-and-isloading-has-been-changed-to-ispending-and-isinitialloading-has-now-been-renamed-to-isloading.
Initially I used `isLoading` when porting to v5, but TypeScript was
not able to narrow the `data` value to non-nil unless I use
`isPending`, so I've gone with `isPending` in the end. As far as I can
tell, it works the same as before.

* v5's dev tools are significantly improved, including the UI, which
lets you choose via a nice drop-down menu where to locate the tool
pop-up. This doesn't play well with our own "dev tools" checkbox UI,
so I've temporarily disabled those `DevOptions`, as the UI for them
will need to be rethought. I think adding another button to the canvas
is probably the way to go, but I'll leave that work for later.

FYI, the migration guide is here, and it informed most of these
changes:

https://tanstack.com/query/latest/docs/framework/react/guides/migrating-to-v5

Signed-off-by: Drew Hess <src@drewhess.com>